KernelSU: Allows manual hooking without using kprobes
Some apps will detect that KernelSU uses kprobes and think that the device
environment is abnormal. Therefore, to avoid being detected, it is best to
follow the official documentation of KernelSU to manually hook, just like
doing it on a traditional kernel.

This commit allows us to manually hook while keeping `CONFIG_KPROBES` enabled.

Notes:
The kernel having kprobes support does not mean that the environment is
abnormal, because GKI has kprobes support by default, and disabling kprobes
support will make it impossible to start the device.

Signed-off-by: Pzqqt <821026875@qq.com>